chore: move messages to app dir
