+++ apps/admin/src/app/messages/unsubscribed/layout.tsx
@@ -1,0 +1,7 @@
+import MessageLayout from "@/components/layouts/MessageLayout";
+
+const Layout = ({ children }) => {
+  return <MessageLayout>{children}</MessageLayout>;
+};
+
+export default Layout;
