+++ apps/admin/src/app/messages/unsubscribed/page.tsx
@@ -1,9 +1,8 @@
+"use client";
 import Head from "next/head";
 import { Result } from "ui";
+
 const Unsubscribed = () => {
-  if (typeof document === "undefined") {
-    return null;
-  }
   const queryParams = new URLSearchParams(document.location.search);
   const msg =
     queryParams.get("msg") || "Your email has been removed from our system";
@@ -21,5 +20,4 @@ const Unsubscribed = () => {
   );
 };
 
-Unsubscribed.isMessage = true;
 export default Unsubscribed;
