+++ apps/admin/src/app/messages/verified/page.tsx
@@ -1,5 +1,6 @@
+"use client";
 import Head from "next/head";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { Button, Result } from "ui";
 
 const Verified = () => {
